fix(notification-response): Add 'FAILED' response
On Trustly debit notification calls (when end-user's account should be
decreased), we have the opportunity to respond with 'FAILED' status
instead of 'OK'. This is needed in deposit cases where it's not possible
to decrease/take back the end-user's credits in the merchant system (in
my case, take back giftcard code).

 - Make it possible to set 'status' to 'FAILED'.
 - Remove unnecessary/unused callback.

https://trustly.com/en/developer/api/#/debit